+++ service/gitlab/gitlab_mr_commit.go
@@ -21,8 +21,9 @@ var _ reviewdog.CommentService = &MergeRequestCommitCommenter{}
 // MergeRequestCommitCommenter is a comment service for GitLab MergeRequest.
 //
 // API:
-//  https://docs.gitlab.com/ce/api/commits.html#post-comment-to-commit
-//  POST /projects/:id/repository/commits/:sha/comments
+//
+//	https://docs.gitlab.com/ce/api/commits.html#post-comment-to-commit
+//	POST /projects/:id/repository/commits/:sha/comments
 type MergeRequestCommitCommenter struct {
 	cli      *gitlab.Client
 	pr       int
@@ -93,10 +94,10 @@ func (g *MergeRequestCommitCommenter) postCommentsForEach(ctx context.Context) e
 				commitID = g.sha
 			}
 			prcomment := &gitlab.PostCommitCommentOptions{
-				Note:     gitlab.String(body),
-				Path:     gitlab.String(loc.GetPath()),
-				Line:     gitlab.Int(lnum),
-				LineType: gitlab.String("new"),
+				Note:     gitlab.Ptr(body),
+				Path:     gitlab.Ptr(loc.GetPath()),
+				Line:     gitlab.Ptr(lnum),
+				LineType: gitlab.Ptr("new"),
 			}
 			_, _, err = g.cli.Commits.PostCommitComment(g.projects, commitID, prcomment, gitlab.WithContext(ctx))
 			return err