+++ service/gitlab/gitlab_mr_discussion.go
@@ -25,8 +25,9 @@ const (
 // MergeRequestDiscussionCommenter is a comment and diff service for GitLab MergeRequest.
 //
 // API:
-//  https://docs.gitlab.com/ee/api/discussions.html#create-new-merge-request-discussion
-//  POST /projects/:id/merge_requests/:merge_request_iid/discussions
+//
+//	https://docs.gitlab.com/ee/api/discussions.html#create-new-merge-request-discussion
+//	POST /projects/:id/merge_requests/:merge_request_iid/discussions
 type MergeRequestDiscussionCommenter struct {
 	cli      *gitlab.Client
 	pr       int
@@ -121,20 +122,20 @@ func (g *MergeRequestDiscussionCommenter) postCommentsForEach(ctx context.Contex
 			continue
 		}
 		eg.Go(func() error {
-			pos := &gitlab.NotePosition{
-				StartSHA:     targetBranch.Commit.ID,
-				HeadSHA:      g.sha,
-				BaseSHA:      targetBranch.Commit.ID,
-				PositionType: "text",
-				NewPath:      loc.GetPath(),
-				NewLine:      lnum,
+			pos := &gitlab.PositionOptions{
+				StartSHA:     gitlab.Ptr(targetBranch.Commit.ID),
+				HeadSHA:      gitlab.Ptr(g.sha),
+				BaseSHA:      gitlab.Ptr(targetBranch.Commit.ID),
+				PositionType: gitlab.Ptr("text"),
+				NewPath:      gitlab.Ptr(loc.GetPath()),
+				NewLine:      gitlab.Ptr(lnum),
 			}
 			if c.Result.OldPath != "" && c.Result.OldLine != 0 {
-				pos.OldPath = c.Result.OldPath
-				pos.OldLine = c.Result.OldLine
+				pos.OldPath = gitlab.Ptr(c.Result.OldPath)
+				pos.OldLine = gitlab.Ptr(c.Result.OldLine)
 			}
 			discussion := &gitlab.CreateMergeRequestDiscussionOptions{
-				Body:     gitlab.String(body),
+				Body:     gitlab.Ptr(body),
 				Position: pos,
 			}
 			_, _, err := g.cli.Discussions.CreateMergeRequestDiscussion(g.projects, g.pr, discussion)